Rewrite line mapping.
Make buffers the primary source of location data. A file is
now a buffer with a path.

The character stream builds the line map inside its attached
buffer.

Introduce a bound location to allow for minimal location data.
Parsers and lexers should be bound to the buffer for the purpose
of diagnostics.